A little more macosx:   enabled PERSIST   made proxyTraphandler   switch on signal type in signal handler   identify refnone But general problem with signal context not revealing register contents. In contact with various Macintosh fora about this problem. So-far confirmed (:-(
Betarun version number: 87
-------------------------------------------------------------------------


git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@13336 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/P/storageblock.c
@@ -682,7 +682,7 @@ static char *DBname(char *path, int checkdir)
       sprintf(dbname, "%s/%s", path, DBNAME);
       return &dbname[0];
     } else {
-      fprintf(output, "DBname: path name length exceeded only %d characters are allowed\n", MAXDBNAMELENGTH - strlen(DBNAME) - 1);
+      fprintf(output, "DBname: path name length exceeded only %d characters are allowed\n", (int)(MAXDBNAMELENGTH - strlen(DBNAME) - 1));
       BetaExit(1);
       return NULL;
     }
@@ -701,7 +701,7 @@ static char *DBname(char *path, int checkdir)
       sprintf(dbname, "%s/%s", path, DBNAME);
       return &dbname[0];
     } else {
-      fprintf(output, "DBname: path name length exceeded only %d characters are allowed\n", MAXDBNAMELENGTH - strlen(DBNAME) - 1);
+      fprintf(output, "DBname: path name length exceeded only %d characters are allowed\n", (int)(MAXDBNAMELENGTH - strlen(DBNAME) - 1));
       BetaExit(1);
       return NULL;
     }
@@ -710,7 +710,7 @@ static char *DBname(char *path, int checkdir)
       sprintf(dbname, "%s.pst", path);
       return &dbname[0];
     } else {
-      fprintf(output, "DBname: path name length exceeded only %d characters are allowed\n", MAXDBNAMELENGTH - strlen(DBNAME) - 1);
+      fprintf(output, "DBname: path name length exceeded only %d characters are allowed\n", (int)(MAXDBNAMELENGTH - strlen(DBNAME) - 1));
       BetaExit(1);
       return NULL;
     }
